+++ profile/website/src/components/index.js
@@ -1,7 +1,7 @@
 export { default as Loader } from './Loader';
 export { default as Navbar } from './Navbar';
 export { default as Footer } from './Footer';
-export { default as Services } from './Services';
+export { default as Skills } from './Skills';
 export { default as Transactions } from './Transactions';
 export { default as Home } from './Home';
 export { default as Smartcontracts} from './Smartcontracts';
